refactor: move goal title rendering logic to another file
